parse TimeType from mysql to LongType in lakeSoul

+++ lakesoul-flink/src/main/java/com/dmetasoul/lakesoul/meta/external/jdbc/JdbcDataTypeConverter.java
@@ -210,7 +210,7 @@ public class JdbcDataTypeConverter {
 //                    }
 //                    return CalendarIntervalType;
 //                }
-                return IntegerType;
+                return LongType;
             case Types.TIMESTAMP:
                 if (adaptiveTimePrecisionMode || adaptiveTimeMicrosecondsPrecisionMode) {
                     if (getTimePrecision(column) <= 3) {
